requirements.txt
Flask==1.1.2
Flask-Dance==3.3.1
Flask-SQLAlchemy==2.5
Flask-Bcrypt==0.7.1
Flask-WTF==0.14.3
Flask-Login==0.5.0
pytest-flask==1.2.0
python-dotenv==0.15.0




